clinic/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
astroid==2.9.3
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coverage==6.3.1
coveralls==3.3.1
cryptography==36.0.1
dnspython==2.2.0
docopt==0.6.2
email-validator==1.1.3
flasgger==0.9.5
Flask==2.0.3
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
mistune==2.0.2
platformdirs==2.5.0
pycparser==2.21
pylint==2.12.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
PyMySQL==1.0.2
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
WTForms==3.0.1
